+++ finalProject/src/main/webapp/WEB-INF/views/common/student_menubar.jsp
@@ -37,7 +37,7 @@
 	    <ul id="nav">
 	        <li><a href="#">홈</a></li>
 	        <li><a href="#">등록/장학</a></li>
-	        <li><a href="#">학사관리</a></li>
+	        <li><a href="infoStudent.st">학사관리</a></li>
 	        <li><a href="counselingList.st">상담관리</a></li>
 	        <li><a href="classListView.st">수강신청</a></li>
 	        <li><a href="classManagement.st">수업관리</a></li>
